requirements.txt
Django==1.6.2
MySQL-python==1.2.5
Pillow==2.5.1
South==0.8.4
argparse==1.2.1
beautifulsoup4==4.3.2
django-classy-tags==0.5
django-crispy-forms==1.4.0
django-jsonfield==0.9.12
django-mobile==0.3.0
django-sekizai==0.7
facebook-sdk==0.4.0
gunicorn==18.0
lxml==3.3.5
pytz==2014.4
wsgiref==0.1.2
